MSVC wants a cast-to-integer-type before dividing by a bool that should promote to int.  Sigh.  Followup to bug 851237, r=windows-redness
UltraBlame original commit: 073556b91f46c2d7387355f9b66cce1614513c7e

+++ js/public/TemplateLib.h
@@ -73,7 +73,7 @@ template <size_t N> struct NBitMask {
     
     
     
-    static const size_t checkPrecondition = 0 / (N < BitSize<size_t>::result);
+    static const size_t checkPrecondition = 0 / size_t(N < BitSize<size_t>::result);
     static const size_t result = (size_t(1) << N) - 1 + checkPrecondition;
 };
 template <> struct NBitMask<BitSize<size_t>::result> {
